% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_dashboard.R
\name{ce_dashboard}
\alias{ce_dashboard}
\title{Run a Cost Effectiveness Interactive Dashboard}
\usage{
ce_dashboard()
}
\value{
Starts a shiny dashboard
}
\description{
This functions runs a dashboard that can be used to plot the results of cost effectiveness analysis.
}
\examples{

## Only run the example if in an interative session
\dontrun{

## Run the dashboard
ce_dashboard()
}
}
