% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_costs.R
\name{check_costs}
\alias{check_costs}
\title{Check the structure of the costs}
\usage{
check_costs(costs = NULL)
}
\arguments{
\item{costs}{A dataframe costs. An example of the format required is \code{ceplotr::example_costs}.}
}
\value{
Nothing is returned
}
\description{
Check the structure of the costs
}
\examples{

}
