% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_health_utility.R
\name{check_health_utility}
\alias{check_health_utility}
\title{Check the structure of the health utilities}
\usage{
check_health_utility(health_utility = NULL)
}
\arguments{
\item{health_utility}{A dataframe health utilities. An example of the format required is \code{ceplotr::example_health_utility}.}
}
\value{
Nothing is returned
}
\description{
Check the structure of the health utilities
}
\examples{

}
