% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_simulation.R
\name{check_simulation}
\alias{check_simulation}
\title{Check the structure of the model simulations}
\usage{
check_simulation(simulations = NULL)
}
\arguments{
\item{simulations}{A dataframe of simulations. An example of the format required is \code{ceplotr::example_simulation}.}
}
\value{
Nothing is returned
}
\description{
Check the structure of the model simulations
}
\examples{

}
