% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Combine Model Simulations, Costs, and Health Utilities}
\usage{
combine_data(simulations = NULL, costs = NULL, health_utility = NULL,
  discounting = NULL)
}
\arguments{
\item{simulations}{A dataframe of simulations. An example of the format required is \code{ceplotr::example_simulation}.}

\item{costs}{A dataframe costs. An example of the format required is \code{ceplotr::example_costs}.}

\item{health_utility}{A dataframe health utilities. An example of the format required is \code{ceplotr::example_health_utility}.}

\item{discounting}{Numeric, defaulting to none.}
}
\value{
A dataframe of combined cost effectiveness data
}
\description{
Combine Model Simulations, Costs, and Health Utilities
}
\examples{

}
