% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_costs.R
\docType{data}
\name{example_costs}
\alias{example_costs}
\title{Example Costs}
\format{A data frame with rows and  variables.
\describe{
}}
\usage{
example_costs
}
\description{
An example of the format that \code{ceplotr} supports for costs
}
\keyword{datasets}
