% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_health_utility.R
\docType{data}
\name{example_health_utility}
\alias{example_health_utility}
\title{Example Health Utilities}
\format{A data frame with rows and  variables.
\describe{
}}
\usage{
example_health_utility
}
\description{
An example of the format that \code{ceplotr} supports for health utilities
}
\keyword{datasets}
