% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_parameters.R
\docType{data}
\name{example_parameters}
\alias{example_parameters}
\title{Example Model Parameters}
\format{A data frame with rows and  variables.
\describe{
}}
\usage{
example_parameters
}
\description{
An example of the format that \code{ceplotr} supports for model parameters
}
\keyword{datasets}
