% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_simulation.R
\docType{data}
\name{example_simulation}
\alias{example_simulation}
\title{Example Model Simulation}
\format{A data frame with rows and  variables.
\describe{
}}
\usage{
example_simulation
}
\description{
An example of the format that \code{ceplotr} supports for model simulations
}
\keyword{datasets}
