% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ceplotr_analysis.R
\name{plot.ceplotr_analysis}
\alias{plot.ceplotr_analysis}
\alias{plot}
\title{Plotting a cost effectiveness analysis}
\usage{
\method{plot}{ceplotr_analysis}(x = NULL, ...)
}
\arguments{
\item{x}{A dataframe created with \code{combine_data}.}

\item{...}{Pass additional arguements to summary function.}
}
\value{
Plots summary measures for a cost effectiveness analysis
}
\description{
\code{plot} method for class "ceplotr_analysis".
}
\examples{



}
\seealso{
plot
}
