% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ce_plane.R
\name{plot_ce_plane}
\alias{plot_ce_plane}
\title{Plot the Cost Effectiveness Plane}
\usage{
plot_ce_plane(ce_data = NULL)
}
\arguments{
\item{ce_data}{A dataframe of cost effectiveness data data as produced by \code{combine_data}.}
}
\value{
A plot of the cost effectiveness plane
}
\description{
Plot the Cost Effectiveness Plane
}
\examples{

}
