% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ceac.R
\name{plot_ceac}
\alias{plot_ceac}
\title{Plot the Cost Effectiveness Acceptability Curve}
\usage{
plot_ceac(ce_data = NULL)
}
\arguments{
\item{ce_data}{A dataframe of cost effectiveness data data as produced by \code{combine_data}.}
}
\value{
A plot of the cost effectiveness acceptability curve
}
\description{
Plot the Cost Effectiveness Acceptability Curve
}
\examples{

}
