% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prcc.R
\name{plot_prcc}
\alias{plot_prcc}
\title{Plot output sensitivty to parameters}
\usage{
plot_prcc(sensitivity = NULL)
}
\arguments{
\item{sensitivity}{A dataframe of sensitivity data as produced by \code{test_prcc}.}
}
\value{
A plot summarising model PRCC sensitivity to parameter changes.
}
\description{
Plot output sensitivty to parameters
}
\examples{

## Code
plot_prcc
}
