% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_willingness.R
\name{plot_willingness}
\alias{plot_willingness}
\title{Plot the Willingness to Pay}
\usage{
plot_willingness(ce_data = NULL)
}
\arguments{
\item{ce_data}{A dataframe of cost effectiveness data data as produced by \code{combine_data}.}
}
\value{
A plot of the willingness to pay
}
\description{
Plot the Willingness to Pay
}
\examples{

}
