% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ceplotr_analysis.R
\name{summary.ceplotr_analysis}
\alias{summary.ceplotr_analysis}
\alias{summary}
\title{Summarising a cost effectiveness analysis}
\usage{
\method{summary}{ceplotr_analysis}(object = NULL, ...)
}
\arguments{
\item{object}{A dataframe created with \code{combine_data}.}

\item{...}{Pass additional arguements to summary function.}
}
\value{
Prints a summary table to the terminal and a summary plot.
}
\description{
\code{summary} method for class "ceplotr_analysis".
}
\examples{



}
\seealso{
summary
}
