% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_prcc.R
\name{test_prcc}
\alias{test_prcc}
\title{Test cost sensitivity using PRCC}
\usage{
test_prcc(df = NULL, obs = NULL, target_time = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{obs}{A character string containing the observation to use to calculate the PRCC.}

\item{target_time}{Numeric, the time at which to estimate the model sensitivity. If not specified
then this will default to the last fitted point that the model has produced output for.}
}
\value{
A data frame containing the names of the parameters in the model, the correlation with the outcome
and the p value of this correlation
}
\description{
Test cost sensitivity using PRCC
}
\examples{

## Code
test_prcc
}
