% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceplotr.R
\docType{package}
\name{ceplotr-package}
\alias{ceplotr}
\alias{ceplotr-package}
\title{ceplotr: Tmp}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Tmp
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.samabbott.co.uk/ceplotr}
  \item \url{https://github.com/seabbs/ceplotr}
  \item Report bugs at \url{https://github.com/seabbs/ceplotr/issues}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk} (0000-0001-8057-8037)

}
\keyword{internal}
